[全般]experimental名前空間から移動､整理 c2xa_config.hpp->config.hpp [通信]パース部分を分離
